Add documentation video link and rename app file as main.rb
