--- app.rb
@@ -1,8 +1,0 @@
-require_relative 'main'
-
-def main
-  app = App.new
-  app.run
-end
-
-main
